% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{abort_config}
\alias{abort_config}
\title{Configuration-related errors}
\usage{
abort_config(message, ...)
}
\arguments{
\item{message}{Error message}

\item{...}{Additional arguments passed to \code{tidyhte_abort()}}
}
\description{
Configuration-related errors
}
\keyword{internal}
