% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{abort_data}
\alias{abort_data}
\title{Data validation errors}
\usage{
abort_data(message, ...)
}
\arguments{
\item{message}{Error message}

\item{...}{Additional arguments passed to \code{tidyhte_abort()}}
}
\description{
Data validation errors
}
\keyword{internal}
