% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{abort_model}
\alias{abort_model}
\title{Model-related errors}
\usage{
abort_model(message, ...)
}
\arguments{
\item{message}{Error message}

\item{...}{Additional arguments passed to \code{tidyhte_abort()}}
}
\description{
Model-related errors
}
\keyword{internal}
