% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{abort_not_implemented}
\alias{abort_not_implemented}
\title{Not implemented errors}
\usage{
abort_not_implemented(message = "Not implemented", ...)
}
\arguments{
\item{message}{Error message}

\item{...}{Additional arguments passed to \code{tidyhte_abort()}}
}
\description{
Not implemented errors
}
\keyword{internal}
