% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{abort_package}
\alias{abort_package}
\title{Package dependency errors}
\usage{
abort_package(message, ...)
}
\arguments{
\item{message}{Error message}

\item{...}{Additional arguments passed to \code{tidyhte_abort()}}
}
\description{
Package dependency errors
}
\keyword{internal}
