% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe-api.R
\name{add_known_propensity_score}
\alias{add_known_propensity_score}
\title{Uses a known propensity score}
\usage{
add_known_propensity_score(hte_cfg, covariate_name)
}
\arguments{
\item{hte_cfg}{\code{HTE_cfg} object to update.}

\item{covariate_name}{Character indicating the name of the covariate
name in the dataframe corresponding to the known propensity score.}
}
\value{
Updated \code{HTE_cfg} object
}
\description{
This replaces the propensity score model with a known value
of the propensity score.
}
\examples{
library("dplyr")
basic_config() \%>\%
   add_known_propensity_score("ps") -> hte_cfg
}
