% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qoi.R
\name{calculate_pcate_quantities}
\alias{calculate_pcate_quantities}
\title{Calculate "partial" CATE estimates}
\usage{
calculate_pcate_quantities(
  full_data,
  .weights,
  .outcome,
  fx_model,
  ...,
  .MCATE_cfg
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
