% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{calculate_vimp}
\alias{calculate_vimp}
\title{Calculate Variable Importance of HTEs}
\usage{
calculate_vimp(.data, pseudo_outcome, ..., .VIMP_cfg)
}
\arguments{
\item{.data}{dataframe}

\item{...}{Unquoted names of covariates to include in the joint effect model. The variable importance
will be calculated for each of these covariates.}

\item{.VIMP_cfg}{A \code{VIMP_cfg} object defining how the joint effect model should be estimated.}

\item{.outcome}{Unquoted name of the pseudo-outcome.}
}
\description{
\code{calculate_vimp} estimates the reduction in (population) $R^2$ from removing a particular moderator
from a model containing all moderators.
}
\references{
Williamson, BD, Gilbert, PB, Carone, M, Simon, N. Nonparametric variable importance
assessment using machine learning techniques. \emph{Biometrics}. 2021; 77: 9-- 22.
\url{https://doi.org/10.1111/biom.13392}
}
