% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{check_data_has_hte_cfg}
\alias{check_data_has_hte_cfg}
\title{Checks that a dataframe has an attached configuration for HTEs}
\usage{
check_data_has_hte_cfg(data)
}
\arguments{
\item{data}{Dataframe of interest.}
}
\value{
Returns NULL. Errors if a problem is discovered.
}
\description{
This helper function ensures that the provided dataframe has
the necessary auxilliary configuration information for HTE
estimation.
}
\keyword{internal}
