% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{check_identifier}
\alias{check_identifier}
\title{Checks that an appropriate identifier has been provided}
\usage{
check_identifier(data, id_col)
}
\arguments{
\item{data}{Dataframe of interest.}

\item{id_col}{Quoted name of identifier column.}
}
\value{
Returns NULL. Errors if a problem is discovered.
}
\description{
This helper function makes a few simple checks to identify obvious
issues with the way provided column of unit identifiers.
}
\keyword{internal}
