% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{check_nuisance_models}
\alias{check_nuisance_models}
\title{Checks that nuisance models have been estimated and exist in the supplied dataset.}
\usage{
check_nuisance_models(data)
}
\arguments{
\item{data}{Dataframe which should have appropriate columns of nuisance function
predictions: \code{.pi_hat}, \code{.mu0_hat}, and \code{.mu1_hat}}
}
\value{
Returns NULL. Errors if a problem is discovered.
}
\description{
This helper function makes a few simple checks to identify obvious
issues with the way that nuisance functions are created and prepared.
}
\keyword{internal}
