% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{check_splits}
\alias{check_splits}
\title{Checks that splits have been properly created.}
\usage{
check_splits(data)
}
\arguments{
\item{data}{Dataframe which should have appropriate \code{.split_id} column.}
}
\value{
Returns NULL. Errors if a problem is discovered.
}
\description{
This helper function makes a few simple checks to identify obvious
issues with the way that splits have been made in the supplied data.
}
\keyword{internal}
