% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{check_weights}
\alias{check_weights}
\title{Checks that an appropriate weighting variable has been provided}
\usage{
check_weights(data, weight_col)
}
\arguments{
\item{data}{Dataframe of interest.}

\item{weight_col}{Quoted name of weights column.}
}
\value{
Returns NULL. Errors if a problem is discovered.
}
\description{
This helper function makes a few simple checks to identify obvious
issues with the weights provided.
}
\keyword{internal}
