% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cfg.R
\name{Diagnostics_cfg}
\alias{Diagnostics_cfg}
\title{Configuration of Model Diagnostics}
\description{
\code{Diagnostics_cfg} is a configuration class for estimating a variety of diagnostics for the models
trained in the course of HTE estimation.
}
\examples{

## ------------------------------------------------
## Method `Diagnostics_cfg$new`
## ------------------------------------------------

Diagnostics_cfg$new(
   outcome = c("SL_risk", "SL_coefs", "MSE"),
   ps = c("SL_risk", "SL_coefs", "AUC")
)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ps}}{Model diagnostics for the propensity score model.}

\item{\code{outcome}}{Model diagnostics for the outcome models.}

\item{\code{effect}}{Model diagnostics for the joint effect model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Diagnostics_cfg$new()}}
\item \href{#method-add}{\code{Diagnostics_cfg$add()}}
\item \href{#method-clone}{\code{Diagnostics_cfg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Diagnostics_cfg} object with specified diagnostics to estimate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Diagnostics_cfg$new(ps = NULL, outcome = NULL, effect = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ps}}{Model diagnostics for the propensity score model.}

\item{\code{outcome}}{Model diagnostics for the outcome models.}

\item{\code{effect}}{Model diagnostics for the joint effect model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Diagnostics_cfg} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Diagnostics_cfg$new(
   outcome = c("SL_risk", "SL_coefs", "MSE"),
   ps = c("SL_risk", "SL_coefs", "AUC")
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Add diagnostics to the \code{Diagnostics_cfg} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Diagnostics_cfg$add(ps = NULL, outcome = NULL, effect = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ps}}{Model diagnostics for the propensity score model.}

\item{\code{outcome}}{Model diagnostics for the outcome models.}

\item{\code{effect}}{Model diagnostics for the joint effect model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Diagnostics_cfg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
