% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{listwise_deletion}
\alias{listwise_deletion}
\title{Removes rows which have missing data on any of the supplied columns.}
\usage{
listwise_deletion(data, ...)
}
\arguments{
\item{data}{The dataset from which to drop cases which are not fully observed.}

\item{...}{Unquoted column names which must be non-missing. Missingness in these
columns will result in dropped observations. Missingness in other columns will not.}
}
\value{
The original data with all observations which are fully observed.
}
\description{
This function removes rows with missingness based on the columns provided.
If rows are dropped, a message is displayed to the user to inform them of this
fact.
}
\keyword{internal}
