% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cfg.R
\name{Model_cfg}
\alias{Model_cfg}
\title{Base Class of Model Configurations}
\description{
\code{Model_cfg} is the base class from which all other model configurations
inherit.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_class}}{The class of the model, required for all classes
which inherit from \code{Model_cfg}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Model_cfg-new}{\code{Model_cfg$new()}}
\item \href{#method-Model_cfg-clone}{\code{Model_cfg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_cfg-new"></a>}}
\if{latex}{\out{\hypertarget{method-Model_cfg-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Model_cfg} object with any necessary parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_cfg$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Model_cfg} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_cfg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Model_cfg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_cfg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
