% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe-api.R
\name{remove_vimp}
\alias{remove_vimp}
\title{Removes variable importance information}
\usage{
remove_vimp(hte_cfg)
}
\arguments{
\item{hte_cfg}{\code{HTE_cfg} object to update.}
}
\value{
Updated \code{HTE_cfg} object
}
\description{
This removes the variable importance quantity of interest
from an \code{HTE_cfg}.
}
\examples{
library("dplyr")
basic_config() \%>\%
   remove_vimp() -> hte_cfg
}
