% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{split_data}
\alias{split_data}
\title{Given a particular split ID, partition the data}
\usage{
split_data(.data, split_id)
}
\arguments{
\item{.data}{dataframe}

\item{split_id}{integer representing the split to construct}
}
\value{
Returns an R6 object \code{HTEFold} with two properties:
\itemize{
\item \code{train} - The split to be used for training the plugin estimates
\item \code{holdout} - The split not used for training
}
}
\description{
This takes a dataset and a split ID and generates two subsets of the
data corresponding to a training set and a holdout.
}
