% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{tidyhte_abort}
\alias{tidyhte_abort}
\title{Throw a tidyhte error}
\usage{
tidyhte_abort(message, class = "general", ...)
}
\arguments{
\item{message}{Error message}

\item{class}{Error class suffix (will be prefixed with "tidyhte_error_")}

\item{...}{Additional arguments passed to \code{rlang::abort()}}
}
\description{
Throw a tidyhte error
}
\keyword{internal}
