% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{attach_config}
\alias{attach_config}
\title{Attach an \code{HTE_cfg} to a dataframe}
\usage{
attach_config(.data, .HTE_cfg)
}
\arguments{
\item{.data}{dataframe}

\item{.HTE_cfg}{\code{HTE_cfg} object representing the full configuration of the HTE analysis.}
}
\description{
This adds a configuration attribute to a dataframe for HTE estimation.
This configuration details the full analysis of HTE that should be performed.
}
