% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe-api.R
\name{basic_config}
\alias{basic_config}
\title{Create a basic config for HTE estimation}
\usage{
basic_config()
}
\value{
\code{HTE_cfg} object
}
\description{
This provides a basic recipe for HTE estimation that can
be extended by providing additional information about models
to be estimated and what quantities of interest should be
returned based on those models. This basic model includes
only linear models for nuisance function estimation.
}
