% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qoi.R
\name{construct_pseudo_outcomes}
\alias{construct_pseudo_outcomes}
\title{Construct Pseudo-outcomes}
\usage{
construct_pseudo_outcomes(.data, y_col, a_col)
}
\arguments{
\item{.data}{dataframe}

\item{y_col}{Unquoted name of outcome variable.}

\item{a_col}{Unquoted name of treatment variable.}
}
\description{
\code{construct_pseudo_outcomes} takes a dataset which has been prepared
with plugin estimators of nuisance parameters and transforms these into
a "pseudo-outcome": an unbiased estimator of the conditional average
treatment effect under exogeneity.
}
