% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qoi.R
\name{construct_pseudo_outcomes}
\alias{construct_pseudo_outcomes}
\title{Construct Pseudo-outcomes}
\usage{
construct_pseudo_outcomes(.data, outcome, treatment)
}
\arguments{
\item{.data}{dataframe (already prepared with \code{attach_config}, \code{make_splits},
and \code{produce_plugin_estimates})}

\item{outcome}{Unquoted name of outcome variable.}

\item{treatment}{Unquoted name of treatment variable.}
}
\description{
\code{construct_pseudo_outcomes} takes a dataset which has been prepared
with plugin estimators of nuisance parameters and transforms these into
a "pseudo-outcome": an unbiased estimator of the conditional average
treatment effect under exogeneity.
}
