% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{estimate_QoI}
\alias{estimate_QoI}
\title{Estimate Quantities of Interest}
\usage{
estimate_QoI(.data, ...)
}
\arguments{
\item{.data}{data frame (already prepared with \code{attach_config}, \code{make_splits},
\code{produce_plugin_estimates} and \code{construct_pseudo_outcomes})}

\item{...}{Unquoted names of moderators to calculate QoIs for.}
}
\description{
\code{estimate_QoI} takes a dataframe already prepared with split IDs,
plugin estimates and pseudo-outcomes and calculates the requested
quantities of interest (QoIs).
}
