% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{estimate_QoI}
\alias{estimate_QoI}
\title{Estimate Quantities of Interest}
\usage{
estimate_QoI(.data, ..., .HTE_cfg = NULL)
}
\arguments{
\item{.data}{data frame}

\item{...}{Unquoted names of moderators to calculate QoIs for.}

\item{.HTE_cfg}{\code{HTE_cfg} object representing the full analysis configuration.}
}
\description{
\code{estimate_QoI} takes a dataframe already prepared with split IDs,
plugin estimates and pseudo-outcomes and calculates the requested
quantities of interest (QoIs).
}
