% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{estimate_QoI}
\alias{estimate_QoI}
\title{Estimate Quantities of Interest}
\usage{
estimate_QoI(data, ...)
}
\arguments{
\item{data}{data frame (already prepared with \code{attach_config}, \code{make_splits},
\code{produce_plugin_estimates} and \code{construct_pseudo_outcomes})}

\item{...}{Unquoted names of moderators to calculate QoIs for.}
}
\description{
\code{estimate_QoI} takes a dataframe already prepared with split IDs,
plugin estimates and pseudo-outcomes and calculates the requested
quantities of interest (QoIs).
}
\examples{
\dontrun{
cfg <- basic_config()
attach_config(data, cfg) \%>\%
make_splits(unitid, .num_splits = 10) \%>\%
produce_plugin_estimates(outcome, treatment, covariate1, covariate2) \%>\%
construct_pseudo_outcomes(outcome, treatment) \%>\%
estimate_QoI(covariate1, covariate2)
}
}
\seealso{
\code{\link[=attach_config]{attach_config()}}, \code{\link[=make_splits]{make_splits()}}, \code{\link[=produce_plugin_estimates]{produce_plugin_estimates()}},
\code{\link[=construct_pseudo_outcomes]{construct_pseudo_outcomes()}},
}
