% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{make_splits}
\alias{make_splits}
\title{Define splits for cross-fitting}
\usage{
make_splits(.data, identifier, ..., .num_splits)
}
\arguments{
\item{.data}{dataframe}

\item{identifier}{Unquoted name of unique identifier column}

\item{...}{variables on which to stratify}

\item{.num_splits}{number of splits to create}
}
\value{
original dataframe with additional \code{.split_id} column
}
\description{
This takes a dataset, a column with a unique identifier and an
arbitrary number of covariates on which to stratify the splits.
It returns the original dataset with an additional column \code{.split_id}
corresponding to an identifier for the split.
}
