% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{produce_plugin_estimates}
\alias{produce_plugin_estimates}
\title{Estimate models of nuisance functions}
\usage{
produce_plugin_estimates(.data, outcome, treatment, ..., .HTE_cfg = NULL)
}
\arguments{
\item{.data}{dataframe}

\item{outcome}{Unquoted name of the outcome variable.}

\item{treatment}{Unquoted name of the treatment variable.}

\item{...}{Unquoted names of covariates to include in the models of the nuisance functions.}

\item{.HTE_cfg}{\code{HTE_cfg} object representing the full configuration of the HTE analysis.}
}
\description{
This takes a dataset with an identified outcome and treatment column along
with any number of covariates and appends three columns to the dataset corresponding
to an estimate of the conditional expectation of treatment (\code{.pi_hat}), along with the
conditional expectation of the control and treatment potential outcome surfaces
(\code{.mu0_hat} and \code{.mu1_hat} respectively).
}
