% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{split_data}
\alias{split_data}
\title{Partition the data into folds}
\usage{
split_data(data, split_id)
}
\arguments{
\item{data}{dataframe}

\item{split_id}{integer representing the split to construct}
}
\value{
Returns an R6 object \code{HTEFold} with three public fields:
\itemize{
\item \code{train} - The split to be used for training the plugin estimates
\item \code{holdout} - The split not used for training
\item \code{in_holdout} - A logical vector indicating for each unit whether they lie in the holdout.
}
}
\description{
This takes a dataset and a split ID and generates two subsets of the
data corresponding to a training set and a holdout.
}
\keyword{internal}
