% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyhte-package.R
\docType{package}
\name{tidyhte-package}
\alias{tidyhte}
\alias{tidyhte-package}
\title{tidyhte: Tidy Estimation of Heterogeneous Treatment Effects}
\description{
Tools for the estimation of heterogeneous treatment effects
    using tidy semantics. Methods are based on the doubly-robust learner
    of Kennedy (n.d.).
}
\details{
The best places to get started with \code{tidyhte} is \code{vignette("example_analysis")} which
walks through a full analysis of HTE on simulated data.
}
\seealso{
The core public-facing functions are \code{make_splits}, \code{produce_plugin_estimates},
\code{construct_pseudo_outcomes} and \code{estimate_QoI}. Configuration is accomplished through \code{HTE_cfg}
in addition to a variety of related classes.
}
\author{
\strong{Maintainer}: Drew Dimmery \email{drew.dimmery@univie.ac.at} (\href{https://orcid.org/0000-0001-9602-6325}{ORCID})

}
\keyword{internal}
