% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cfg.R
\name{VIMP_cfg}
\alias{VIMP_cfg}
\title{Configuration of Variable Importance}
\description{
\code{VIMP_cfg} is a configuration class for estimating a variable importance measure across all moderators.
This provides a meaningful measure of which moderators explain the most of the CATE surface.
}
\examples{

## ------------------------------------------------
## Method `VIMP_cfg$new`
## ------------------------------------------------

VIMP_cfg$new(model_cfg = SLLearner_cfg$new("SL.glm"))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_cfg}}{A \code{Model_cfg} object indicating how to fit the second level effect
regression (joint across all moderators).}

\item{\code{estimand}}{String indicating the estimand to target.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VIMP_cfg$new()}}
\item \href{#method-clone}{\code{VIMP_cfg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{VIMP_cfg} object with specified model configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VIMP_cfg$new(model_cfg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_cfg}}{A \code{Model_cfg} object indicating how to fit the second level effect
regression (joint across all moderators).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VIMP_cfg} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{VIMP_cfg$new(model_cfg = SLLearner_cfg$new("SL.glm"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VIMP_cfg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
